+++ src/components/P2P/handleProtocolCommands.ts
@@ -10,7 +10,7 @@ import { P2PCommandResponse } from '../../@types'
 import { P2P_CONSOLE_LOGGER } from './index.js'
 
 import { findDDO } from '../core/handlers/ddoHandler.js'
-import { NonceHandler } from '../core/handlers/nonceHandler.js'
+import { NonceHandler } from '../core/handlers/handler.js'
 import { FeesHandler } from '../core/handlers/feesHandler.js'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
 import { StatusHandler } from '../core/handlers/statusHandler.js'
