+++ src/components/core/handlers/getDdoHandler.ts
@@ -1,4 +1,4 @@
-import { Handler } from './aHandler.js'
+import { Handler } from './handler.js'
 import { GetDdoCommand } from '../../../utils/constants.js'
 import { Database } from '../../database/index.js'
 import { P2PCommandResponse } from '../../../@types/index.js'
